// frontend/src/api/generated/api.ts
export interface SensorData {
    'id'?: number;
    'sensorId': number;
    'timestamp': string;
    'value': number;
}
// ...
export class SensorApi extends BaseAPI {
    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiDataGet(options?: RawAxiosRequestConfig) {
        return SensorApiFp(this.configuration).apiDataGet(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {SensorData} [sensorData] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiDataPost(sensorData?: SensorData, options?: RawAxiosRequestConfig) {
        return SensorApiFp(this.configuration).apiDataPost(sensorData, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiSensorsSummaryGet(options?: RawAxiosRequestConfig) {
        return SensorApiFp(this.configuration).apiSensorsSummaryGet(options).then((request) => request(this.axios, this.basePath));
    }
}